requirements.txt
requests
beautifulsoup4
lxml
schedule
PyYAML
python-dotenv
selenium
webdriver-manager